requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-dotenv==1.0.1
pydantic==2.9.2
numpy==1.26.4
Pillow==10.4.0
# Optional: for future extensions
# ortools==9.10.4067
# trimesh==4.4.9
# open3d==0.18.0
numpy
